rename: 예외처리 로직이 있기 때문에 DTO -> VO로 변경
